+++ src/mattermostautodriver/endpoints/search.py
@@ -18,3 +18,18 @@ class Search(Base):
 
         """
         return self.client.post(f"/api/v4/teams/{team_id}/files/search", data=data)
+
+    def search_files(self, data):
+        """Search files across the teams of the current user
+
+        terms: The search terms as entered by the user. To search for files from a user include ``from:someusername``, using a user's username. To search in a specific channel include ``in:somechannel``, using the channel name (not the display name). To search for specific extensions include ``ext:extension``.
+        is_or_search: Set to true if an Or search should be performed vs an And search.
+        time_zone_offset: Offset from UTC of user timezone for date searches.
+        include_deleted_channels: Set to true if deleted channels should be included in the search. (archived channels)
+        page: The page to select. (Only works with Elasticsearch)
+        per_page: The number of posts per page. (Only works with Elasticsearch)
+
+        `Read in Mattermost API docs (search - SearchFiles) <https://api.mattermost.com/#tag/search/operation/SearchFiles>`_
+
+        """
+        return self.client.post("""/api/v4/files/search""", data=data)