+++ src/mattermostautodriver/endpoints/timeline.py
@@ -12,4 +12,4 @@ class Timeline(Base):
         `Read in Mattermost API docs (Timeline - removeTimelineEvent) <https://api.mattermost.com/#tag/Timeline/operation/removeTimelineEvent>`_
 
         """
-        return self.client.delete(f"/plugins/playbooks/api/v0/runs/{id}/timeline/{event_id}/")
+        return self.client.delete(f"/plugins/playbooks/api/v0/runs/{id}/timeline/{event_id}")